requirements.txt
Flask>=2.0.0
python-dotenv>=1.0.0
google-genai>=0.1.0
